Update package.json
Because of the fact that you're using import, specifying the type as a module will allow it to be imported easily into node.js projects. Since this is not common.js (Doesn't use `require()`), this does not integrate very well with any node.js project (In my specific case next.js).